EleProtect_App/requirements.txt
streamlit==1.32.0
biopython==1.81
pandas==2.2.3
openpyxl==3.1.2
python-docx==0.8.11
streamlit-lottie==0.0.5  # optional; used for animation if available
joblib==1.3.2